requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
filelock==3.12.4
gdown==4.7.1
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.0
numpy==1.24.1
Pillow==9.3.0
polars==0.19.3
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
soupsieve==2.5
sympy==1.12
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tqdm==4.66.1
typing_extensions==4.4.0
urllib3==2.0.5
